Co-locate game events with command files
