+++ MooSharp/Commands/SystemMessageEvent.cs
@@ -1,0 +1,12 @@
+using MooSharp.Messaging;
+
+namespace MooSharp;
+
+public record SystemMessageEvent(string Message) : IGameEvent;
+
+public class SystemMessageEventFormatter : IGameEventFormatter<SystemMessageEvent>
+{
+    public string FormatForActor(SystemMessageEvent gameEvent) => gameEvent.Message;
+
+    public string FormatForObserver(SystemMessageEvent gameEvent) => gameEvent.Message;
+}
